% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani-aai-methods.R
\name{ani_filter}
\alias{ani_filter}
\title{Title}
\usage{
ani_filter(
  ani,
  checkm,
  gtdbtk,
  comp = 50,
  cont = 20,
  hetero = 20,
  species = NULL
)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Title
}
