% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract_genomelist}
\alias{extract_genomelist}
\title{Title}
\usage{
extract_genomelist(ortho, delete_genomes = 0, num_ortho = 10)
}
\arguments{
\item{delete_genomes}{}
}
\value{

}
\description{
Title
}
\examples{

}
