% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract_ortho}
\alias{extract_ortho}
\title{Title}
\usage{
extract_ortho(ortho, freq = NULL, genome_list = NULL, delete_genomes = 0)
}
\arguments{
\item{delete_genomes}{}
}
\value{

}
\description{
Title
}
\examples{

}
