% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_aai}
\alias{plot_aai}
\title{Title}
\usage{
plot_aai(
  polaris,
  plot_type = c("heatmap", "distribution"),
  subcluster = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
