% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_gtdbtk}
\alias{plot_gtdbtk}
\title{Barplot using gtdbtk data with metadata}
\usage{
plot_gtdbtk(
  gtdbtk,
  metadata = NULL,
  category = NULL,
  type = c("single", "ani", "pplacer"),
  taxonomy = c("domain", "phylum", "class", "order", "family", "genus", "species"),
  colors = colors,
  ...
)
}
\value{

}
\description{
Barplot using gtdbtk data with metadata
}
