% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_ortho}
\alias{plot_ortho}
\title{Plot Ortyholog and phylogenetic tree}
\usage{
plot_ortho(
  ortho,
  tree,
  metadata,
  num_ortho = 10,
  delete_genomes = 0,
  tip_lab = T
)
}
\arguments{
\item{delete_genomes}{}
}
\value{

}
\description{
Orthofinderとphylogenetic treeをまとめて書く
}
\examples{

}
