% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_eggnog}
\alias{read_eggnog}
\title{Eggnong annotationの読み込み}
\usage{
read_eggnog(file, name = NULL, extension = ".annotations")
}
\arguments{
\item{file}{eggnongの結果、eggnog annotation}
}
\value{

}
\description{
Eggnogデータを操作しやすい形で読み込む
おそらく、懸念点としては、各cell内でカンマで
分けられてるデータをリスト化すること
}
