% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_eggnogs}
\alias{read_eggnogs}
\title{Eggnogデータのまとめて読み込み}
\usage{
read_eggnogs(
  eggnog,
  dir = FALSE,
  extension = "annotations",
  names = NULL,
  tools = c("eggnog", "atlas")
)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
Eggnong-mapperの結果をまとめて読み込む
}
\examples{


}
