% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_fastas}
\alias{read_fastas}
\title{Read fasta files from specific directory}
\usage{
read_fastas(
  fastas,
  dir = FALSE,
  type = c("DNA", "AA"),
  extension = "fna",
  names = NULL,
  ...
)
}
\arguments{
\item{fastas}{fastaのファイルパスのリストもしくはディレクトリ}

\item{dir}{(default: FALSE)}

\item{type}{配列のタイプ, DNAかAAか}

\item{extension}{fastaファイルの拡張子。読み込みの際に指定された拡張子はファイル名から削除されて、ゲノム名となる}

\item{names}{ゲノム の名前を自分で指定したい場合に指定する}
}
\value{
DNAStringSetと名前のリスト
}
\description{
fastaファイルをまとめて読み込むための関数
}
\examples{
#fasta_dir <- system.file("inst/extdata/Rickettsiales/fasta", package = "polaris")
#fasta_list <- read_fastas(fasta_dir, dir = TRUE, type = "DNA")

# fastaのファイル一覧取得。こちらでも動く
# fasta_filelist <- list.files(path = fastas, pattern = NULL, full.names = T)
# fasta_list <- read_fastas(fasta_filelist)
}
