% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_gff}
\alias{read_gff}
\title{Read gff files from specific directory}
\usage{
read_gff(file, tibble = FALSE, extension = "gff")
}
\arguments{
\item{file}{gffファイルのパス}

\item{tibble}{(default: FALSE) 出力をgffclassではなくtibbleにする}
}
\value{
gff gffのクラス
}
\description{
gffファイルを一つ読み込むための関数
}
\examples{

}
