% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_gffs}
\alias{read_gffs}
\title{複数のgffファイルの読み込み}
\usage{
read_gffs(files, dir = FALSE, extension = "gff", names = NULL)
}
\value{

}
\description{
複数のgffファイルを一つ一つ読み込んで統合することが少々煩雑なので、
最初から複数ファイルを読み込むための関数を提供する
}
