% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_gtdbtk}
\alias{read_gtdbtk}
\title{GTDBtkのtsvを読み込む}
\usage{
read_gtdbtk(file)
}
\arguments{
\item{file}{gtdbtk.bac120.summary.tsv}
}
\value{

}
\description{
GTDBtkのtsvを読み込む
}
\examples{
#file <- system.file("inst/extdata/Rickettsiales/gtdbtk/gtdbtk.bac120.summary.tsv", package = "polaris")
#gtdb <- read_gtdbtk(file)

# polaris \%<>\%
#  add_polaris(gtdb)
#
}
