% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_interpro}
\alias{read_interpro}
\title{interproscanの読み込み}
\usage{
read_interpro(file, type = c("tsv", "json", "xml"))
}
\arguments{
\item{file}{interproscanファイル}

\item{type}{ファイルのタイプ3パターン}
}
\value{

}
\description{
基本的にはtsvを読み込むことにする
複数のゲノムがある場合には、
}
