% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_interpros}
\alias{read_interpros}
\title{Title}
\usage{
read_interpros(
  files,
  dir = FALSE,
  type = c("tsv", "json", "xml"),
  names = NULL
)
}
\value{

}
\description{
Title
}
\examples{


}
