% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-methods.R
\name{read_ortho}
\alias{read_ortho}
\title{Orthofinderの結果を読み込む}
\usage{
read_ortho(ortho_file)
}
\arguments{
\item{ortho_file}{orthofinderの結果、orthofroups.tsv}
}
\value{
ortho_class ortho
}
\description{
Orthofinderの結果を読み込む。出力としては3つある。
1つ目はどの遺伝子とどのOrthogroupが対応しているか
2つ目はどのOrthogroupがどのゲノムにどれくらいあるか
3つ目はシングルコピーオルソログはどれか
}
