% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_gtdbtk}
\alias{plot_gtdbtk}
\title{Title}
\usage{
plot_gtdbtk(
  gtdbtk,
  metadata,
  type = c("single", "ani", "pplacer"),
  taxonomy = c("domain", "phylum", "class", "order", "family", "genus", "species"),
  ...
)
}
\value{

}
\description{
Title
}
